test chart config with yq

+++ demo-oai.sh
@@ -1011,6 +1011,11 @@ configure-gnb() {
     for nf in oai-gnb oai-gnb-fhi-72 oai-du oai-cu oai-cu-cp oai-cu-up; do
 	VALUES="${OAI5G_RAN}/${nf}/values.yaml"
 	echo "***** nf: $nf, $VALUES"
+
+	if [[ ! -f "$VALUES" ]]; then
+            echo "Skipping $nf: file not found"
+            continue
+	fi	
 	cp "$VALUES" "${OAI5G_RAN}/${nf}/values.yaml.orig"
 
 	apply-gnb-values-yq "$VALUES"
